Moved the lib/ to the src/ directory.
This is in preparation of eventually adding typescript.
