+++ Makefile
@@ -11,9 +11,9 @@ test:
 
 .PHONY: lint
 lint:
-	node_modules/.bin/eslint lib/
+	node_modules/.bin/eslint src/
 
-dist/ketting.min.js: lib/*.js lib/*/*.js
+dist/ketting.min.js: src/*.js src/*/*.js
 	mkdir -p dist
 	node_modules/.bin/webpack \
 		--optimize-minimize \
