+++ src/followable-promise.js
