+++ test/integration/put.js
@@ -1,5 +1,5 @@
-const Ketting = require('../../lib/ketting');
-const Resource = require('../../lib/resource');
+const Ketting = require('../../src/ketting');
+const Resource = require('../../src/resource');
 const expect = require('chai').expect;
 
 describe('Issuing a PUT request', async () => {
