+++ test/unit/ketting.js
@@ -1,6 +1,6 @@
-const Ketting = require('../../lib/ketting');
-const Html = require('../../lib/representor/html');
-const Hal = require('../../lib/representor/hal');
+const Ketting = require('../../src/ketting');
+const Html = require('../../src/representor/html');
+const Hal = require('../../src/representor/hal');
 const expect = require('chai').expect;
 
 describe('Ketting', () => {
